+++ client-sdk/src/errors.ts
@@ -1,6 +1,11 @@
+interface ErrorOptions {
+  cause?: Error;
+}
+
 export class RTAPError extends Error {
   constructor(message: string, options?: ErrorOptions) {
-    super(message, options);
+    super(message);
     this.name = 'RTAPError';
+    this.cause = options?.cause;
   }
 }